+++ LibreHardwareMonitorLib/PawnIo/LpcIO.cs
@@ -1,17 +1,10 @@
-﻿using System;
+﻿namespace LibreHardwareMonitor.PawnIo;
 
-namespace LibreHardwareMonitor.PawnIo;
-
-internal class LpcIO
+internal class LpcIo
 {
-    readonly long[] _singleArgArray = new long[1];
-    readonly long[] _doubleArgArray = new long[2];
-    readonly PawnIo _pawnIO;
-
-    public LpcIO()
-    {
-        _pawnIO = PawnIo.LoadModuleFromResource(typeof(LpcIO).Assembly, $"{nameof(LibreHardwareMonitor)}.Resources.PawnIO.LpcIO.bin");
-    }
+    private readonly long[] _doubleArgArray = new long[2];
+    private readonly PawnIo _pawnIO = PawnIo.LoadModuleFromResource(typeof(LpcIo).Assembly, $"{nameof(LibreHardwareMonitor)}.Resources.PawnIO.LpcIO.bin");
+    private readonly long[] _singleArgArray = new long[1];
 
     public void SelectSlot(int slot)
     {
